# .github/workflows/swagger.yaml
openapi: 3.0.3
info:
  title: CargoHUB API
  description: API voor het beheren van klanten, orders, inventaris, locaties, zendingen, en transfers binnen CargoHUB.
  version: 1.0.0
servers:
  - url: https://localhost:5072/cargohub
    description: Productieomgeving
  - url: https://sandbox.api.cargohub.com
    description: Sandbox omgeving
paths:
  /clients/{client_id}:
    get:
      summary: Haal een specifieke klant op
      parameters:
        - name: client_id
          in: path
          required: true
          description: ID van de klant
          schema:
            type: string
      responses:
        '200':
          description: Klant gevonden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Client'
        '404':
          description: Klant niet gevonden
    put:
      summary: Werk een specifieke klant bij
      operationId: updateClient
      parameters:
        - name: client_id
          in: path
          required: true
          description: ID van de klant
          schema:
            type: string
      requestBody:
        description: Klantgegevens om bij te werken
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Client'
      responses:
        '200':
          description: Klant bijgewerkt
        '404':
          description: Klant niet gevonden
    delete:
      summary: Verwijder een specifieke klant
      operationId: deleteClient
      parameters:
        - name: client_id
          in: path
          required: true
          description: ID van de klant
          schema:
            type: string
      responses:
        '204':
          description: Klant verwijderd
        '404':
          description: Klant niet gevonden
  /clients:
    get:
      summary: Haal alle klanten op
      description: Haalt een lijst van alle klanten op
      operationId: readClients
      responses:
        '200':
          description: Klanten gevonden
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Client'
        '404':
          description: Geen klanten gevonden
    
    post:
      summary: Voeg een nieuwe klant bij
      operationId: postClient
      requestBody:
        description: Voegt een nieuwe klant toe aan de database
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Client'
        required: true
      responses:
        '200':
          description: Klant toegevoegd
          content:
            application/json:
              schema:
                type: string
                example: Klant toegevoegd
        '400':
          description: Klant niet kunnen toevoegen
          content:
            application/json:
              schema:
                type: string
                example: Klant niet kunnen toevoegen
  /clients/{client_id}/orders:
    get:
      summary: Haalt alle orders van een client op
      description: Haalt een lijst op van alle orders van een client
      operationId: readClientOrders
      parameters:
        - name: client_id
          in: path
          required: true
          description: ID van de klant om orders op te halen
          schema:
            type: integer
      responses:
        '200':
          description: Klant met orders gevonden
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
        '404':
          description: Geen klant met orders gevonden
  /inventories/{inventory_id}:
    get:
      summary: Read an inventory by ID
      parameters:
        - name: inventory_id
          in: path
          required: true
          description: The ID of the inventory to retrieve
          schema:
            type: integer
      responses:
        '200':
          description: Inventory found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Inventory'
        '404':
          description: Inventory not found
          content:
            text/plain:
              schema:
                type: string
                example: "No such inventory with Id: {inventory_id}"

    put:
      summary: Update an existing inventory
      parameters:
        - name: inventory_id
          in: path
          required: true
          description: The ID of the inventory to update
          schema:
            type: integer
      requestBody:
        description: Inventory object that needs to be updated
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Inventory'
      responses:
        '200':
          description: Inventory updated successfully
          content:
            text/plain:
              schema:
                type: string
                example: "Inventory updated successfully."
        '400':
          description: Failed to update inventory
          content:
            text/plain:
              schema:
                type: string
                example: "Failed to update inventory."

    delete:
      summary: Delete an inventory by ID
      parameters:
        - name: inventory_id
          in: path
          required: true
          description: The ID of the inventory to delete
          schema:
            type: integer
      responses:
        '200':
          description: Inventory deleted successfully
          content:
            text/plain:
              schema:
                type: string
                example: "Inventory deleted successfully."
        '400':
          description: Failed to delete inventory
          content:
            text/plain:
              schema:
                type: string
                example: "Failed to delete inventory."
  /inventories:
    get:
      summary: Retrieve all inventories
      responses:
        '200':
          description: Inventories found
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Inventory'
        '404':
          description: No inventories found
          content:
            text/plain:
              schema:
                type: string
                example: "No inventories found"
    post:
      summary: Create a new inventory
      requestBody:
        description: Inventory object to be created
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Inventory'
      responses:
        '200':
          description: Inventory created successfully
          content:
            text/plain:
              schema:
                type: string
                example: "Inventory created successfully."
        '400':
          description: Failed to create inventory
          content:
            text/plain:
              schema:
                type: string
                example: "Failed to create inventory."
...
  /items/{item_id}/inventory:
    get:
      summary: Get inventories for a specific item
      parameters:
        - name: item_id
          in: path
          required: true
          description: The ID of the item to get inventories for.
          schema:
            type: string
      responses:
        200:
          description: Inventories found
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Inventory'
        404:
          description: No inventories found for item
          content:
            text/plain:
              schema:
                type: string
                example: "No inventories found for item with Id: {item_id}"
...
components:
  schemas:
    Client:
      type: object
      properties:
        id:
          type: integer
          description: Unieke identificatie van de klant
        name:
          type: string
          description: Naam van de klant
        address:
          type: string
          description: Adres van de klant
        city:
          type: string
          description: Woonplaats van de klant
        zip_code:
          type: string
          description: Zip code van de klant
        province:
          type: string
          description: Provincie van de klant
        country:
          type: string
          description: Land van de klant
        contact_name:
          type: string
          description: Naam van de klant
        contact_phone:
          type: string
          description: Telefoonnummer van de klant
        contact_email:
          type: string
          format: email
          description: Email adres van de klant
        email:
          type: string
          format: email
          description: E-mailadres van de klant
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Inventory:
      type: object
      properties:
        id:
          type: integer
          description: Unieke identificatie van het inventaris
        item_id:
          type: string
          description: Item identificatie
        description:
          type: string
          description: Beschrijving van het item
        item_reference:
          type: string
          description: Referentie van het item
        locations:
          type: array
          items:
            type: integer
          description: Lijst van locaties
        total_on_hand:
          type: integer
          description: Aantal op voorraad
        total_expected:
          type: integer
          description: Aantal verwachte items
        total_ordered:
          type: integer
          description: Aantal bestelde items
        total_allocated:
          type: integer
          description: Aantal toegewezen items
        total_available:
          type: integer
          description: Aantal beschikbare items
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Order:
      type: object
      properties:
        id:
          type: integer
          description: Order ID
        source_id:
          type: integer
          description: Bron ID
        order_date:
          type: string
          description: Datum van de bestelling
        request_date:
          type: string
          description: Verzoekdatum
        reference:
          type: string
          description: Referentie van de order
        reference_extra:
          type: string
          description: Extra referentie
        order_status:
          type: string
          description: Status van de bestelling
        notes:
          type: string
          description: Notities voor de bestelling
        shipping_notes:
          type: string
          description: Verzendnotities
        picking_notes:
          type: string
          description: Picking notities
        warehouse_id:
          type: integer
          description: Magazijn ID
        ship_to:
          type: integer
          description: Verzendadres ID
        bill_to:
          type: integer
          description: Factuuradres ID
        shipment_id:
          type: integer
          description: Zending ID
        total_amount:
          type: number
          format: float
          description: Totaalbedrag van de bestelling
        total_discount:
          type: number
          format: float
          description: Totale korting van de bestelling
        total_tax:
          type: number
          format: float
          description: Totale belasting van de bestelling
        total_surcharge:
          type: number
          format: float
          description: Totale toeslag van de bestelling
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        items:
          type: array
          items:
            type: object
            properties:
              item_id:
                type: string
              quantity:
                type: integer
...